refactor(MeshIterator): use CellType enum instead of std::uint8_t
When iterating through mesh cells, the iterator return the type as a CellType enum instead of a std::uint8_t. It allow to avoid a static_cast to compare the type to the enum.

+++ libs/core/fwData/test/tu/src/MeshTest.cpp
@@ -925,7 +925,7 @@ void MeshTest::iteratorTest()
         size_t count = 0;
         for (; it != itEnd; ++it)
         {
-            *it->type   = static_cast<std::uint8_t>(::fwData::Mesh::CellType::QUAD);
+            *it->type   = ::fwData::Mesh::CellType::QUAD;
             *it->offset = 4*count;
             if (it != itEnd-1)
             {
